Add debug pipeline page
